requirements.txt
Flask==1.1.2

pyrebase4==4.5.0requests==2.23.0
scikit-learn==0.22.1
seaborn==0.10.1
setuptools>=41.0.0
gunicorn==19.7.1
